% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_image.R
\name{as_image}
\alias{as_image}
\title{Transform matrix into image (Transformar matriz em imagem).}
\usage{
as_image(img)
}
\arguments{
\item{img}{:Object image  (Objeto com uma imagem).}
}
\value{
Returns an object of type image (Retorna um objeto do tipo imagem).
}
\description{
This function transform matrix into image  (Esta funcao transforma matriz em imagem).
}
\examples{
im1=read_image(example_image(2),plot=TRUE)
r=gray_scale(im = im1,method = "r")
R=as_image(r)
}
\seealso{
\code{\link{info_image}} , \code{\link{read_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
