% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_palette.R
\name{calibration_palette}
\alias{calibration_palette}
\title{Function to create a color palette for image color calibration (Funcao para criar uma paleta de cores para calibracao das cores de imagens).}
\usage{
calibration_palette(n=4,px=500,file=NULL,plot=TRUE)
}
\arguments{
\item{n}{:Number of color tones to be created (n.n.n). Numero de tons de cores que se deseja criar (n.n.n).}

\item{px}{:Number of rows and columns in the image (Numero de linhas e de colunas na imagem).}

\item{file}{:Address of the file to be created (Endereco do arquivo a ser criado).}

\item{plot}{:Plot the image  (Plota a imagem).}
}
\value{
Returns an image with a color palette (Retorna um imagem com uma paleta de cores).
}
\description{
This function creates a palette for color calibration of images.(Esta funcao cria uma paleta para calibracao das cores de imagens).
}
\examples{
im=calibration_palette(n=4,px = 500)
im=calibration_palette(n=3,px = 100)
}
\seealso{
\code{\link{info_image}} , \code{\link{read_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
