% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_image.R
\name{compare_image}
\alias{compare_image}
\title{Function to compare image sizes  (Funcao para comparar os tamanhos das imagens).}
\usage{
compare_image(imgs=NULL,filesnames=NULL,path = NULL)
}
\arguments{
\item{imgs}{:List object containing the images  (Objeto do tipo lista contendo as imagens).}

\item{filesnames}{:Images names (Nomes das imagens).}

\item{path}{:Path files  (Endereco das pastas).}
}
\value{
Return images size (Retorna o tamanho das imagens).
}
\description{
This function compares the size of images  (Esta funcao compara o tamanho das imagens).
}
\examples{
#Carregando imagens de exemplo
im1=read_image(example_image(1))
im2=read_image(example_image(2))
im3=read_image(example_image(3))

#Comparando o tamanho das imagens
compare_image(list(im1,im2,im3))
}
\seealso{
\code{\link{info_image}} , \code{\link{read_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
