% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_image.R
\name{contour_image}
\alias{contour_image}
\title{Contorno da imagem segmentada (Outline of the segmented image.)}
\usage{
contour_image(img,plot=TRUE)
}
\arguments{
\item{img}{Este objeto deve conter uma imagem contida em uma matriz binaria.}

\item{plot}{Indica se sera apresentada (TRUE) ou nao (FALSE) (default) a
imagem segmentada (Indicates whether the segmented image will be
displayed (TRUE) or not (FALSE) (default)).}
}
\value{
Retorna o contorno de objetos em uma imagem binaria (Returns the outline of objects in a binary image ).
}
\description{
Esta funcao retorna o contorno da imagem segmentada
(This function returns the outline of the segmented image).
}
\examples{
end=example_image(3)
im=read_image(end,TRUE)
b=gray_scale(im = im,method = "g",plot = TRUE)
m=segmentation(b,threshold = "otsu",selectHigher = FALSE,fillHull = TRUE,plot=TRUE)
contour_image(m)


}
\seealso{
\code{\link{edit_image}} ,  \code{\link{edit_imageGUI}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
