% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_color.R
\name{convert_color}
\alias{convert_color}
\title{Convert colors (Converter cores).}
\usage{
convert_color(color)
}
\arguments{
\item{color}{A vector with 3 values (r, g and b) or a data.frame with 3 columns (r, g and b)
(Um vetor com 3 valores(r, g e b) ou um data.frame com 3 colunas (r, g e b)).}
}
\value{
Displays colors in RGB for LAB, HEX and XYZ
(Apresenta as cores em RGB para LAB, HEX e XYZ)
}
\description{
This function converts RGB colors to LAB, HEX and XYZ
(Esta funcao converte cores RGB para LAB, HEX e XYZ).
}
\examples{
#Valores de RGB
color=c(0.15,0.16,0.17)
convert_color(color)
color=data.frame(R=runif(10),G=runif(10),B=runif(10))
color
convert_color(color)
}
\seealso{
\code{\link{pick_color}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
