% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_imageGUI.R
\name{edit_imageGUI}
\alias{edit_imageGUI}
\title{This function changes the brightness, contrast and makes the gamma adjustment
in the image (Esta funcao permite mudar o brilho, contraste e fazer um ajuste
gamma na imagem).}
\usage{
edit_imageGUI(im,Verbose=TRUE)
}
\arguments{
\item{im}{:This object must contain an image in EBImage format (Este
objeto deve conter uma imagem no formato do EBImage).}

\item{Verbose}{:Logical value. Show the configurations  (Valor logico.
Mostra as configuracoes).}
}
\value{
Returns a edited image (Retorna uma imagem editada).
}
\description{
This function changes the brightness, contrast and makes the
  gamma adjustment in the image (Esta funcao permite mudar o brilho,
  contraste e fazer um ajuste gamma na imagem).
}
\examples{
\dontrun{
#library(ExpImage)
#Carregar imagem de exemplo
im=read_image(example_image(2),plot=TRUE)

##Editar imagem
im2=edit_imageGUI(im)
}
}
\seealso{
\code{\link{edit_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
