% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpImage-package.R
\docType{package}
\name{ExpImage-package}
\alias{ExpImage-package}
\alias{ExpImage}
\title{ExpImage: Experimental Image Analysis Tools}
\description{
This package was
 made for image analysis in researches. There are functions
 associated with image editing, segmentation, and obtaining
 biometric measurements (Este pacote foi idealizado para para a analise de
  imagens em pesquisas. Ha funcoes associadas a edicao de imagens,
  segmentacao, e obtencao de medidas biometricas).
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>
}
\seealso{
\code{\link{crop_image}}, \code{\link{edit_image}},
 \code{\link{example_image}},\code{\link{extract_pixels}},
  \code{\link{gray_scale}}, \code{\link{join_image}},
  \code{\link{mask_pixels}}
}
\author{
Alcinei Mistico Azevedo <alcineimistico@hotmail.com>
}
