% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_bands.R
\name{join_bands}
\alias{join_bands}
\title{Function to compare image sizes  (Funcao para comparar os tamanhos das imagens).}
\usage{
join_bands(imgs=NULL,filesnames=NULL,path = NULL)
}
\arguments{
\item{imgs}{:List object containing the images  (Objeto do tipo lista contendo as imagens).}

\item{filesnames}{:Images names (Nomes das imagens).}

\item{path}{:Path files  (Endereco das pastas).}
}
\value{
Return images size (Retorna o tamanho das imagens).
}
\description{
This function compares the size of images  (Esta funcao compara o tamanho das imagens).
}
\examples{
#Carregando imagens de exemplo
im1=read_image(example_image(2),plot=TRUE)
r=gray_scale(im = im1,method = "r")
g=gray_scale(im = im1,method = "g")
b=gray_scale(im = im1,method = "b")

im2=join_bands(imgs = list(r,g,b))
plot_image(im2)
}
\seealso{
\code{\link{info_image}} , \code{\link{read_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
