% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_raster.R
\name{mask_raster}
\alias{mask_raster}
\title{Creates a mask over the background or foreground in a raster image (Cria uma mascara sobre o
background ou foreground em uma imagem no formato raster)}
\usage{
mask_raster(Class,legend,col,main,plot=TRUE)
}
\arguments{
\item{Class}{Raster object with pixels grouped into classes
(Objeto do tipo raster com os pixels agrupados em classes).}

\item{legend}{Vector with the names corresponding to each class (Vetor os nomes correspondentes a cada classe).}

\item{col}{Vector with the color names corresponding to each class (Vetor os nomes das cores correspondentes a cada classe).}

\item{main}{Titulo da figura (Figure title)}

\item{plot}{Logic value, if TRUE the figure will be shown (Valor logico, se TRUE a figura sera apresentada)}
}
\value{
Returns an array (retorna uma matriz).
}
\description{
Function to create a mask above the background or foreground.
(Funcao para criar uma mascara sobre o background ou foreground).
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
