% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_image.R
\name{normalize_image}
\alias{normalize_image}
\title{Create a histogram from image bands
(Criar um histrograma a partir das bandas das imagens)}
\usage{
normalize_image(im,inf=0,max=1)
}
\arguments{
\item{im}{:Este objeto deve conter uma imagem (This object must contain an image ).}

\item{inf}{: Numero em que o pixel de menor valor devera ter
(Number in which the lowest value pixel should have).}

\item{max}{: Numero em que o pixel de maior valor devera ter
(Number in which the pixel with the highest value should have).}
}
\value{
Retorna histogramas a partir das bandas (Return histograms from the bands of the images).
}
\description{
This function allows you to create histograms from the bands of the images (Esta funcao permite criar histogramas a partir das bandas das imagens).
}
\examples{
\donttest{
end=example_image(6)
im=read_image(end,plot=TRUE)
histogram_image(im,layout = 1)
histogram_image(im,layout = 2)
histogram_image(im,layout = 3)

im2=normalize_image(im)
plot_image(im2)
histogram_image(im2,layout = 3)
}
}
\seealso{
\code{\link{segmentation_logit}}
}
