% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_circle.R
\name{pick_circle}
\alias{pick_circle}
\title{Segmentation with the shape of a circle by clicking on the image
(Segmentacao com o formato de um circulo clicando em na imagem)}
\usage{
pick_circle(im,num.points=NULL, col="red")
}
\arguments{
\item{im}{:This object must contain an image in EBImage format (Este
objeto deve conter uma imagem no formato do EBImage).}

\item{num.points}{:Maximum number of points to be used to delimit the object
(Número maximo de pontos a ser utilizado para delimitar o objeto).}

\item{col}{: Color of points and lines that delimit the object
(Cor dos pontos e linhas que delimitarão o objeto)}
}
\value{
Returns an image with pixel values equal to 1(white) for the foreground and 0 (black)
 for the foreground (Retorna uma imagem com valores de pixel igual a 1( branco) para o foreground
  e 0 (preto) para o foreground).
}
\description{
With this function it is possible to make a segmentation with the shape of a
 circle by clicking on the image (Com essa funcao e possivel fazer uma segmentacao com o
 formato de um circulo clicando em na imagem).
}
\examples{

\dontrun{
im=read_image("https://raw.githubusercontent.com/AlcineiAzevedo/Files_ExpImage/main/PlacaPetri.jpg",
plot=TRUE)
seg=pick_circle(im)
im2=extract_pixels(im,target = seg,plot=TRUE)

}
}
\seealso{
\code{\link{segmentation}} , \code{\link{segmentation_logit}}, \code{\link{segmentation_logitGUI}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
