% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_color.R
\name{pick_color}
\alias{pick_color}
\title{Selecionar os valores de RGB em um pixel (Selecting RGB values in a pixel).}
\usage{
pick_color(im)
}
\arguments{
\item{im}{Este objeto deve conter uma imagem no formato do EBImage (This
object must contain an image in EBImage format ).}
}
\value{
Esta funcao retorna o valor de R, G e B no pixel selecionado.
(This function returns the value of R, G and B at the selected pixel ).
}
\description{
Esta funcao retorna o valor de R, G e B no pixel selecionado.
(This function returns the value of R, G and B at the selected pixel ).
}
\examples{
\donttest{
#library(ExpImage)
#library(EBImage)
#Carregar imagem de exemplo
im=read_image(example_image(2),plot=TRUE)
pick_color(im)
}
}
\seealso{
\code{\link{segmentation_logit}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
