% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_count.R
\name{pick_count}
\alias{pick_count}
\title{Count objects (Contar objetos).}
\usage{
pick_count(im,cex=1,col="red")
}
\arguments{
\item{im}{:Image in ExpImage format (Imagem no formato do ExpImage).}

\item{cex}{:Font size (Tamanho da fonte)}

\item{col}{: Font color (Cor da fonte).}
}
\value{
Returns the number of objects in an image (Retorna o numero de objetos em uma imagem).
}
\description{
Count objects by clicking on the image (Contar objetos clicando na imagem)
}
\examples{
\dontrun{
im=read_image(example_image(3),plot=TRUE)
pick_count(im)
}

}
\seealso{
\code{\link{pick_length}} , \code{\link{pick_resolution}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
