% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_length.R
\name{pick_length}
\alias{pick_length}
\title{Measure objects in the image on click (Medir objetos na imagem ao clicar).}
\usage{
pick_length(im,dpi=NULL,NumPoints=2,col="red")
}
\arguments{
\item{im}{:Image in ExpImage format (Imagem no formato do ExpImage).}

\item{dpi}{:Dots Per Inch - Number of pixels per inch in the image  (Numero de pixels por polegadas na imagem)}

\item{NumPoints}{: Number of points on the segment to be measured (Numero de pontos no segmento que se deseja medir).}

\item{col}{: Font color (Cor da fonte).}
}
\value{
Returns the length of a segment in an image on click (Retorna o comprimento de um segmento em uma imagem ao clicar).
}
\description{
Measure objects in the image on click (Medir objetos na imagem ao clicar).
}
\examples{
\dontrun{
im=read_image(example_image(3),plot=TRUE)
dpi=pick_resolution(im,centimeters=8.5)
dpi$dpi
pick_length(im,dpi = dpi$dpi)

}

}
\seealso{
\code{\link{pick_color}} , \code{\link{pick_resolution}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
