% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_image.R
\name{plot_image}
\alias{plot_image}
\title{Esta funcao corta a imagem retirando as laterais nao desejadas. (This function
removes unwanted sides from the images.)}
\usage{
plot_image(im)
}
\arguments{
\item{im}{Este objeto deve conter uma imagem no formato do EBImage ou na forma
de uma matriz, no caso de imagem em escala de cinza (This
 object must contain an image in EBImage format).}
}
\value{
Apresenta a imagem contida no objeto im.
}
\description{
Esta funcao permite cortar a imagem (This function allows you to
 crop the image).
}
\examples{

#library(ExpImage)
#library(EBImage)
#Carregar imagem de exemplo
im=read_image(example_image(2),plot=TRUE)
plot_image(im)

}
\seealso{
\code{\link{read_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
