% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2image.R
\name{raster2image}
\alias{raster2image}
\title{Convert a image in raster format to ExpImage format (Converte uma imagem no formato raster para o formato ExpImage)}
\usage{
raster2image(im,Bands=NULL,norma=1)
}
\arguments{
\item{im}{: Image in a raster format (imagem do formato raster)}

\item{Bands}{: Vector with the values of desired layers (Vetor contendo os numeros correspondentes as camadas desejadas)}

\item{norma}{: Value used as the denominator to have values pixels whith values between 0 and 1 (Valor utilizado como divisor para ter pixels com valores entre 0 e 1 )}
}
\value{
Returns an raster image in ExpImage format (Retorna um imagem no formato do ExpImage).
}
\description{
This function converts a image in raster format to ExpImage format (Esta funcao converte uma imagem no formato raster para o formato ExpImage)
}
\seealso{
\code{\link{info_image}} , \code{\link{read_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
