% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_image.R
\name{rotate_image}
\alias{rotate_image}
\title{Esta funcao rotaciona a imagem. (This function rotates the image .)}
\usage{
rotate_image(im,angle=NULL,BGcolor=c(0,0,0),plot=TRUE)
}
\arguments{
\item{im}{Este objeto deve conter uma imagem no formato do EBImage (This
object must contain an image in EBImage format ).}

\item{angle}{Valor em graus (Degree value).}

\item{BGcolor}{Vetor com os valores que preencherao o background (Vector with
the values that will fill the background).}

\item{plot}{Indica se sera apresentada (TRUE) ou nao (FALSE) (default) a
imagem segmentada (Indicates whether the segmented image will be
displayed (TRUE) or not (FALSE) (default)).}
}
\value{
Retorna uma imagem rotacionada (Returns a rotated image).
}
\description{
Esta funcao permite rotacionar a imagem (This function allows
rotate the image).
}
\examples{
#library(ExpImage)
#Carregar imagem de exemplo
im=read_image(example_image(2),plot=TRUE)

##rotacionar a imagem
im2=rotate_image(im,angle=80)
im2=rotate_image(im,angle=60,BGcolor=c(1,1,1))
\dontrun{
im2=rotate_image(im)

}
}
\seealso{
\code{\link{edit_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
