% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_script.R
\name{setwd_script}
\alias{setwd_script}
\title{Definir o diretório do script working directory  (Set the script Directory as working directory).}
\usage{
setwd_script()
}
\value{
Esta funcao indica o diretorio de trabalho a pasta onde o script esta salvo
(This function indicates the working directory as the folder where the script is saved).
}
\description{
Esta funcao define como o endereco da pasta onde esta o scritp atual como o diretorio de trabalho.
(This function set the directory of current script as working directory).
}
