% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonize_image.R
\name{skeletonize_image}
\alias{skeletonize_image}
\title{Function for skeletonization in images (Esta funcao faz a skeletonizacao em
imagens).}
\usage{
skeletonize_image(x,plot=FALSE)
}
\arguments{
\item{x}{:Este objeto deve conter uma imagem em uma matriz binaria.}

\item{plot}{:Se forigual a TRUE a imagem sera plotada.
#'@author Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)}
}
\value{
Retorna uma imagem com os objetos skeletonizados.
}
\description{
Esta funcao permite fazer a skeletonizacao em imagens.
}
\examples{
im=read_image(example_image(10),plot=TRUE)
im2=segmentation(im@.Data[,,1],plot = TRUE)
T1=skeletonize_image(im2,plot = TRUE)
T2=thinning_image(im2,plot = TRUE)
}
\seealso{
\code{\link{thinning_image}}
}
