% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Square_Image.R
\name{Square_Image}
\alias{Square_Image}
\title{Transform the square image (Transforme a imagem quadrada).}
\usage{
Square_Image(im,r=1,g=1,b=1,plot=FALSE)
}
\arguments{
\item{im}{:Image in ExpImage format (Imagem no formato do ExpImage).}

\item{r}{:Value from 0 to 1 corresponding to the Red band (Valor de 0 a 1 correspondente a banda de Vermelho).}

\item{g}{:Value from 0 to 1 corresponding to the Green band (Valor de 0 a 1 correspondente a banda de Verde).}

\item{b}{:Value from 0 to 1 corresponding to the Blue band (Valor de 0 a 1 correspondente a banda de Azul).}

\item{plot}{:TRUE if you want to see the image resized (TRUE se desejar ver a imagem redimensionada).}
}
\value{
Returns the resolution in DPI of an image. (Retorna a resolução em uma imagem).
}
\description{
Transforms the image square by placing borders at the edges
(Transforma a imagem quadrada colocando bordas nas margens).
}
\examples{
im=read_image(example_image(10),plot=TRUE)
Square_Image(im,r=0,g=1,b=0,plot=TRUE)

}
\seealso{
\code{\link{pick_color}} , \code{\link{pick_length}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
