% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_image.R
\name{thinning_image}
\alias{thinning_image}
\title{Esta funcao faz o thinning em imagens.}
\usage{
thinning_image(x,plot=F)
}
\arguments{
\item{x}{:Este objeto deve conter uma imagem em uma matriz binaria.}

\item{plot}{:Se forigual a TRUE a imagem sera plotada.}
}
\value{
Retorna uma imagem com o thinning dos objetos.
}
\description{
Esta funcao permite fazer o thinning em imagens.
}
\examples{
\dontrun{
library(ExpImage.pt)
}
}
\seealso{
\code{\link{thinning_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
