% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif_info.R
\name{tif_info}
\alias{tif_info}
\title{Image information in tif format (Informacoes das imagens no formato tif).}
\usage{
tif_info(img=NULL,file=NULL)
}
\arguments{
\item{img}{Este objeto deve conter uma imagem no formato do EBImage ou na forma
de uma matriz, no caso de imagem em escala de cinza (This
 object must contain an image in EBImage format).}

\item{file}{:File name or image address (Nome do arquivo ou endereco da imagem).}
}
\value{
Function to get information from images in tif format
(Funcao para obter informacoes das imagens no formato tif).
}
\description{
Function to get information from images in tif format
(Funcao para obter informacoes das imagens no formato tif).
}
\examples{
########################################################
###' Abrindo o endereco de bandas de imagens de satelite
########################################################
end1=example_image(14) #Banda de azul

########################################################
###' Abrindo bandas de imagens de satelite
B1=read_image(end1,plot=TRUE)
########################################################
###' Informacao das bandas de imagens de satelite
 tif_info(B1)

}
\seealso{
\code{\link{crop_image}} , \code{\link{edit_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
