% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_image.R
\name{as_image}
\alias{as_image}
\title{Transform matrix into image (Transformar matriz em imagem).}
\usage{
as_image(img)
}
\arguments{
\item{img}{:Object image  (Objeto com uma imagem).}
}
\value{
Returns an object of type image (Retorna um objeto do tipo imagem).
}
\description{
This function transform matrix into image  (Esta funcao transforma matriz em imagem).
}
\examples{
m=matrix(sort(runif(10000)),ncol=100,nrow=100)
m
info_image(m)
im=as_image(m)
im
info_image(im)
plot_image(m)
}
\seealso{
\code{\link{info_image}} , \code{\link{read_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
