% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_image.R
\name{crop_image}
\alias{crop_image}
\title{Esta funcao corta a imagem retirando as laterais nao desejadas. (This function
removes unwanted sides from the images.)}
\usage{
crop_image(im,w=NULL,h=NULL,segmentation=NULL,plot=TRUE,verbose=FALSE)
}
\arguments{
\item{im}{Este objeto deve conter uma imagem no formato do EBImage (This
object must contain an image in EBImage format).}

\item{w}{Deve ser um vetor contendo os numeros das colunas que permanecerao na
imagem (It must be a vector containing the column numbers that will remain
in the image).}

\item{h}{Deve ser um vetor contendo os numeros das linhas que permanecerao na
imagem (It must be a vector containing the numbers of the lines that will
remain in the image ).}

\item{segmentation}{matrix binaria obtida por uma segmentacao
(Binary matrix obtained of a segmentation)}

\item{plot}{Indica se sera apresentada (TRUE) ou nao (FALSE) (default) a
imagem segmentada (Indicates whether the segmented image will be
displayed (TRUE) or not (FALSE) (default)).}

\item{verbose}{Indica se sera apresentada (TRUE) ou nao (FALSE) (default) os
pontos de corte (Indicates whether the segmented image will be
 displayed (TRUE) or not (FALSE) (default) the points crop).}
}
\value{
Retorna uma imagem cortada, apresentando apenas os  pixels
 selecionados (Returns a cropped image showing only selected pixels).
}
\description{
Esta funcao permite cortar a imagem (This function allows you to
 crop the image).
}
\examples{
#library(ExpImage)
#Carregar imagem de exemplo
im=read_image(example_image(2),plot=TRUE)


##Cortar Imagem

im3=crop_image(im,w =286:421,h=242:332,plot = TRUE)

# intefacie grafica
\dontrun{
im2=crop_image(im)
}

}
\seealso{
\code{\link{edit_image}} ,  \code{\link{edit_imageGUI}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
