% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_image.R
\name{example_image}
\alias{example_image}
\title{anco de imagens utilizadas como exemplo.}
\usage{
example_image(ex)
}
\arguments{
\item{ex}{:numero do exemplo.}
}
\value{
Retorna o endereco das imagens de exempo.
}
\description{
Apresenta endereco de imagens utilizadas no arquivo de exemplo.
}
\examples{
\dontrun{
example_image(1)
example_image(2)
}
}
