% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwd_script.R
\name{getwd_script}
\alias{getwd_script}
\title{Obter o diretorio onde esta o script. (Get the script directory.)}
\usage{
getwd_script()
}
\description{
Esta funcao retorna o edereco da pasta onde esta o scritp atual
(This function returns the directory of current script ).
}
