% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gray_scale.R
\name{gray_scale}
\alias{gray_scale}
\title{uncao para a obtencao de uma imagem em escala de cinza a partir de uma imagem
colorida}
\usage{
gray_scale(im,method="r",plot=F)
}
\arguments{
\item{im}{:Este objeto deve conter uma imagem no formato do EBImage.}

\item{method}{: Este objeto indica o metodo para a obtencao da escala de
cinza.:\cr
 "r" = extrair a banda de vermelho\cr
  "g" = extrair a banda de
verde\cr "b" = extrair a banda de azul\cr
 "rg" = considera a media da banda
de vermelho e verde: (r+g)/2\cr
 "rb" = considera a media da banda de
vermelho e azul: (r+b)/2\cr
 "gb" = considera a media da banda de verde e
azul: (g+b)/2\cr
 "gbb" = considera a media das 3 bandas: (r+g+b)/3\cr}

\item{plot}{:Indica se sera apresentada (TRUE) ou nao (FALSE) (default) a
imagem segmentada.}
}
\value{
Retorna uma imagem com um texto sobreposto a cada objeto na imagem
}
\description{
Esta funcao permite a obtencao de uma imagem em escala de cinza
  a partir de uma imagem colorida.
}
\examples{
\dontrun{
#Carregar imagem de exemplo
im=readImage(example_image(2))
##mostrar imagem
plot(im)


#Extraindo as camadas R, G e B
r=gray_scale(im,method = "r",plot=T)
g=gray_scale(im,method = "g",plot=T)
b=gray_scale(im,method = "b",plot=T)
}
}
\seealso{
\code{\link{segmentation_logit}}
}
