% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_image.R
\name{histogram_image}
\alias{histogram_image}
\title{Create a histogram from image bands
(Criar um histrograma a partir das bandas das imagens)}
\usage{
histogram_image(im,layout=1)
}
\arguments{
\item{im}{:Este objeto deve conter uma imagem (This object must contain an image ).}

\item{layout}{: Valor numerico variando entre 1 e 3 para se ter diferentes layouts
(Numeric value ranging between 1 and 3 to have different layouts).}
}
\value{
Retorna histogramas a partir das bandas (Return histograms from the bands of the images).
}
\description{
This function allows you to create histograms from the bands of the images (Esta funcao permite criar histogramas a partir das bandas das imagens).
}
\examples{
end=example_image(6)
im=read_image(end,plot=TRUE)
histogram_image(im,layout = 1)
histogram_image(im,layout = 2)
histogram_image(im,layout = 3)
}
\seealso{
\code{\link{segmentation_logit}}
}
