% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_image.R
\name{join_image}
\alias{join_image}
\title{Join images(Juntar imagens)}
\usage{
join_image(im1=NULL,im2=NULL,im3=NULL,im4=NULL,im5=NULL,bk=c(1,1,1),
ncol=NULL,normalize=FALSE,plot=TRUE)
}
\arguments{
\item{im1}{:Object containing an array, image in EBImage format or list of images
(Objeto contendo um array, imagem, ou lista de imagens no formato do EBImage).}

\item{im2}{:Object containing an array, image in EBImage format or list of images
(Objeto contendo um array, imagem, ou lista de imagens no formato do EBImage).}

\item{im3}{:Object containing an array, image in EBImage format or list of images
(Objeto contendo um array, imagem, ou lista de imagens no formato do EBImage).}

\item{im4}{:Object containing an array, image in EBImage format or list of images
(Objeto contendo um array, imagem, ou lista de imagens no formato do EBImage).}

\item{im5}{:Object containing an array, image in EBImage format or list of images
(Objeto contendo um array, imagem, ou lista de imagens no formato do EBImage).}

\item{bk}{:Vector white rgb values for background(Vetor contendo os valores de rgb
que serao considerados no background)}

\item{ncol}{:Number of columns where images will appear in the chart (Numero de colunas em que as imagens aparecerao no grafico)}

\item{normalize}{:Logical value indicating whether the image needs to be normalized
(Valor logico indicando se a imagem precisa ser nomalizada)}

\item{plot}{:Logical value. Indicates whether the image will be displayed (TRUE) or not
(FALSE) (default) (Indica se sera apresentada (TRUE) ou nao (FALSE)
(default) a imagem segmentada).}
}
\value{
Return multiple images joined (Retorna a uniao de varias imagens).
}
\description{
This function joins images by placing them side by side (Esta funcao junta imagens colocando uma do lado da outra).
}
\examples{
\donttest{
im1=read_image(example_image(2),plot=TRUE)

##Aumentar brilho
im2=edit_image(im1,brightness = 0.1)

#Aumentar contraste
im3=edit_image(im1,contrast = 1.2)

#Aumentar gamma
im4=edit_image(im1,gamma  = 1.1)

#Alterando brilho, contraste e gamma
im5=edit_image(im3,brightness = 0.1,contrast = 1.7,gamma  = 1.2)

#Juntando imagens em um objeto raster e plotando
join_image(im1,im2,im3,im4,im5)

#Juntando imagens em um objeto raster e plotando com uma lista
List1=list(im1,im2,im3,im4,im5)
join_image(List1)



im1b=read_image(example_image(7),plot=TRUE)

##Aumentar brilho
im2b=edit_image(im1b,brightness = 0.1)

#Aumentar contraste
im3b=edit_image(im1b,contrast = 1.2)

#Aumentar gamma
im4b=edit_image(im1b,gamma  = 1.1)

#Alterando brilho, contraste e gamma
im5b=edit_image(im3b,brightness = 0.1,contrast = 1.7,gamma  = 1.2)

#Juntando imagens em um objeto raster e plotando
join_image(im1b,im2b,im3b,im4b,im5b)

#Juntando imagens em um objeto raster e plotando com uma lista
List2=list(im1b,im2b,im3b,im4b,im5b)
join_image(List2)

#Juntando imagens presentes em duas listas
join_image(List1,List2)
}
}
\seealso{
\code{\link{crop_image}}, \code{\link{edit_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
