% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_color.R
\name{pick_color}
\alias{pick_color}
\title{Selecionar os valores de RGB em pixels (Selecting RGB values in pixels).}
\usage{
pick_color(im, matrix = NULL)
}
\arguments{
\item{im}{Este objeto deve conter uma imagem no formato do EBImage (This
object must contain an image in EBImage format ).}

\item{matrix}{Deve ser uma matriz com 3 colunas correspondentes aos canais
R, G e B. Os novos valores selecionados serao adicionados a essa matriz.
(It must be a matrix with 3 columns corresponding to the R, G and B channels.
 The new selected values will be added to this matrix).}
}
\value{
Esta funcao retorna o valor de R, G e B no pixel selecionado.
(This function returns the value of R, G and B at the selected pixel ).
}
\description{
Esta funcao retorna o valor de R, G e B no pixel selecionado.
(This function returns the value of R, G and B at the selected pixel ).
}
\examples{
\dontrun{
#library(ExpImage)
#library(EBImage)
#Carregar imagem de exemplo
im=read_image(example_image(2),plot=TRUE)
pick_color(im)
}

}
\seealso{
\code{\link{segmentation_logit}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
