% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indexes.R
\name{plot_indexes}
\alias{plot_indexes}
\title{Function to plot multiple indices (funcao para plotar varios indices)}
\usage{
plot_indexes(im)
}
\arguments{
\item{im}{:This object must contain an image in EBImage format (Este
objeto deve conter uma imagem no formato do EBImage).}
}
\description{
Function to plot multiple indices (funcao para plotar varios
 indices).
}
\examples{
\donttest{
#Carregar imagem de exemplo
im=read_image(example_image(2))
plot_indexes(im)
}
}
\seealso{
\code{\link{gray_scale}}
}
