% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image.R
\name{read_image}
\alias{read_image}
\title{Function to import an image (Funcao para importa uma imagem).}
\usage{
read_image(file,plot=FALSE)
}
\arguments{
\item{file}{:Nome do arquivo ou endereco da imagem.}

\item{plot}{:Indica se sera apresentada (TRUE) ou nao (FALSE) (default) a
imagem}
}
\value{
Importa uma imagem.
}
\description{
Esta funcao importa uma imagem.
}
\examples{
#Carregar imagem de exemplo
im=read_image(example_image(1),plot=TRUE)
}
\seealso{
\code{\link{crop_image}} , \code{\link{edit_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
