% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_logit.R
\name{segmentation_logit}
\alias{segmentation_logit}
\title{Funcao para a segmentacao de imagens.}
\usage{
segmentation_logit(im,foreground,background,return="image",
   sample=2000, fillHull=TRUE,TargetPixels="all",plot=TRUE)
}
\arguments{
\item{im}{:Este objeto deve ser obrigatoriamente uma imagem colorida (RGB)
no formato do EBImage).}

\item{foreground}{: Deve ser uma imagem correspondende a paleta de cores do
objeto que se pretende segmentar. Caso haja mais de uma paleta de cores,
suas reespectivas imagens devem ser colocadas dentro de um objeto do tipo
list. Cada paleta de cor desve estar no formato de imagens do EBImage.}

\item{background}{: Deve ser uma imagem correspondende a paleta de cores
com os tons do fundo. Caso haja mais de uma paleta de cores, suas
reespectivas imagens devem ser colocadas dentro de um objeto do tipo list.
Cada paleta de cor desve estar no formato de imagens do EBImage.}

\item{return}{Texto indicando o objeto a ser exportado pela funcao. Para
este argumento podemos considerar: \cr
  "image" = sera exportada uma matriz referente a imagem segmentada.\cr
  "model" = Sera exportado o modelo para a predicao.\cr}

\item{sample}{: Deve ser um valor numerico indicando quantos pixels dos
imagens do foreground e do background serao utilizados no ajuste do modelo
logit. O valor a ser escolhido deve ser inferior ou igual ao numero de
pixels contidos nas paletas de cores.}

\item{fillHull}{:Este argumento deve receber a palavra TRUE quando se
pretende desconsiderar valores vazios dentro do foreground, caso contrario
FALSE.}

\item{TargetPixels}{:Quando se pretende segmentar todos os pixeis da imagem
deve considerar a palavra "all" (Default). Se a segmentacao deva ser feita
apenas para um conjunto de pixels, estes devem ser apresentados em uma
matriz contendo o valor 1 para os pixels de interesse e 0 para os demais.}

\item{plot}{:Indica se sera apresentada (TRUE) ou nao (FALSE) (default) a
imagem segmentada.}
}
\value{
Imagem segmentada
}
\description{
Esta funcao possibilita a segmentacao de imagens por meio de do
 ajuste de um modelo linear generalizado com a funcao logit de ligacao.
}
\examples{
\dontrun{
im=read_image(example_image(7),plot=TRUE)
segmentation_logitGUI(im)  }
}
\seealso{
\code{\link{glm}} ,\code{\link{segmentation}}
}
\author{
Alcinei Mistico Azevedo (Instituto de Ciencias Agrarias da UFMG)
}
