% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_image.R
\name{thinning_image}
\alias{thinning_image}
\title{This function makes the thinning images(Esta funcao faz o thinning em imagens).}
\usage{
thinning_image(x,plot=FALSE)
}
\arguments{
\item{x}{:Este objeto deve conter uma imagem em uma matriz binaria.}

\item{plot}{:Se forigual a TRUE a imagem sera plotada.}
}
\value{
Retorna uma imagem com o thinning dos objetos.
}
\description{
Esta funcao permite fazer o thinning em imagens.
}
\examples{
im=read_image(example_image(10),plot=TRUE)
im2=segmentation(im@.Data[,,1],plot = TRUE)
T1=skeletonize_image(im2,plot = TRUE)
T2=thinning_image(im2,plot = TRUE)
}
\seealso{
\code{\link{thinning_image}}
}
\author{
Alcinei Mistico Azevedo (Instituto de ciencias agrarias da UFMG)
}
